refactor(mobile): dedupe TransactionTable date format via client-safe util
formatTxDate re-spelled the Format.DATE options that formatDate(Format.DATE)
already owns. The dup existed because formatDate's default pulls in the
server-only env (DATE_LOCALE) and TransactionTable is a client component.

Extract the client-safe primitives (Format enum, options, formatDateWithLocale)
into utils/formatDateCore.ts with no server-only import; formatDate re-uses and
re-exports them, and TransactionTable formats via the core util — single source
of truth for the option set, no server coupling in the client bundle.
