+++ utils/formatDate.ts
@@ -1,33 +1,14 @@
+import { Format, formatDateWithLocale } from './formatDateCore';
 import getDefaultDateLocale from './getDefaultDateLocale';
 
-export enum Format {
-  DATE,
-  MONTH_YEAR,
-  SHORT_MONTH_YEAR,
-}
+// Re-exported so existing `import formatDate, { Format } from '@/utils/formatDate'`
+// call sites keep working; client components should import the client-safe
+// primitives from './formatDateCore' directly.
+export { Format, formatDateWithLocale } from './formatDateCore';
 
-const formatOptions: Record<Format, Intl.DateTimeFormatOptions> = {
-  [Format.DATE]: {
-    year: 'numeric',
-    month: '2-digit',
-    day: '2-digit',
-  },
-  [Format.MONTH_YEAR]: {
-    year: 'numeric',
-    month: 'long',
-  },
-  [Format.SHORT_MONTH_YEAR]: {
-    year: 'numeric',
-    month: 'short',
-  },
-};
-
-const formatDate = (date: string, format: Format) => {
-  return new Date(date).toLocaleDateString(
-    getDefaultDateLocale(),
-    formatOptions[format]
-  );
-};
+/** Format a date using the server-configured default locale (`DATE_LOCALE`). */
+const formatDate = (date: string, format: Format) =>
+  formatDateWithLocale(date, format, getDefaultDateLocale());
 
 export default formatDate;
